requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
PyMySQL==1.1.1
cryptography==43.0.3
redis==5.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.12
google-auth==2.35.0
requests==2.32.3
email-validator==2.2.0
